move GeometricShapeFactoryExt to nitrite package namespace
