+++ nitrite-spatial/src/main/java/org/dizitart/no2/spatial/NearFilter.java
@@ -19,7 +19,6 @@ package org.dizitart.no2.spatial;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.Point;
-import org.locationtech.jts.util.GeometricShapeFactoryExt;
 
 /**
  * @since 4.0
